pyproject.toml
[project]
name = "cav4-mock-server"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi-cli>=0.0.7",
    "fastapi[standard]>=0.115.12",
    "jinja2>=3.1.6",
    "pyjwt>=2.10.1",
]
